+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 ### 2020-03-01
 - Added script to look for a kafka message by key or by the value of a field in case of JSON/AVRO messages
+
+### 2020-03-06
+- Added script to generate an access token for Google APIs